// src/app/api/ui/models/trade-ext-customer-info-excel.interface.ts
// @ts-ignore
export interface TradeExtCustomerInfoExcel {
  bankName: string;
  bicBank: string;
  checkingAccount: string;
  contractEndDate: string;
  contractNumber: string;
  contractProlongationNoChanges: string;
  contractResponsibleEmail: string;
  contractResponsibleName: string;
  contractResponsiblePatronimic: string;
  contractResponsiblePhoneNumber: string;
  contractResponsiblePosition: string;
  contractResponsibleSurname: string;
  contractStartDate: string;
  edmOperator: string;
  faxNumber: string;
  signatoryAuthorityDocument: string;
  signatoryName: string;
  signatoryPatronimic: string;
  signatoryPosition: string;
  signatorySurname: string;
}